requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
distro==1.9.0
frozenlist==1.8.0
greenlet==3.3.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.0
langchain-core==1.2.5
langchain-ollama==1.0.1
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.1
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
numpy==2.4.0
ollama==0.6.1
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
Pygments==2.19.2
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
sniffio==1.3.1
SQLAlchemy==2.0.45
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
